Update Bootstrap Layout
Bootstrap doesn't need the offsets. This causes the page to display incorrectly. The offsets are useful when centering or when swapping two columns on different screen sizes, but on this occasion, there is already a column with width 2 and another with 10, so you can't offset by another 2 as that equals 14. Let me know if they were there for another purpose.